fixing...

+++ src/main/kotlin/net/ccbluex/liquidbounce/utils/entity/SimulatedPlayer.kt
@@ -73,7 +73,7 @@ class SimulatedPlayer(
     var pitch: Float,
     private var sprinting: Boolean,
 
-    var fallDistance: Float,
+    var fallDistance: Double,
     private var jumpingCooldown: Int,
     private var isJumping: Boolean,
     private var isFallFlying: Boolean,
@@ -305,7 +305,7 @@ class SimulatedPlayer(
             var k: Double
             var e: Vec3d = this.velocity
             if (e.y > -0.5) {
-                fallDistance = 1.0f
+                fallDistance = 1.0
             }
             val vec3d3 = this.getRotationVector()
             val f: Float = this.pitch * (Math.PI.toFloat() / 180)
@@ -510,7 +510,7 @@ class SimulatedPlayer(
     }
 
     private fun onLanding() {
-        this.fallDistance = 0.0f
+        this.fallDistance = 0.0
     }
 
     fun jump() {
